pyproject.toml
[tool.poetry]
name = "delta-loader"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.6.1"
pyspark = "^3.3.0"
delta-spark = "^2.1.0"
boto3 = "^1.24.70"
rich = "^12.5.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
chispa = "^0.9.2"
ipython = "^8.5.0"
jupyterlab = "^3.4.6"
localstack = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
